+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- Refundmanager is now disabled for orders in Authorized state, it is not possible to refund an order if nothing was caputred yet
 
 ### Fixed
 - Fixed webhook race condition